requirements.txt
asgiref==3.5.2
black==22.10.0
CacheControl==0.12.11
cachetools==5.2.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cryptography==38.0.3
Django==4.1.3
django-environ==0.9.0
django-phonenumber-field==7.0.0
djangorestframework==3.14.0
ffmpeg-python==0.2.0
firebase-admin==6.0.1
future==0.18.2
google-api-core==2.10.2
google-api-python-client==2.65.0
google-auth==2.14.1
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-firestore==2.7.2
google-cloud-secret-manager==2.12.6
google-cloud-storage==2.6.0
google-crc32c==1.5.0
google-resumable-media==2.4.0
googleapis-common-protos==1.56.4
grpc-google-iam-v1==0.12.4
grpcio==1.50.0
grpcio-status==1.50.0
gunicorn==20.1.0
httplib2==0.21.0
idna==3.4
msgpack==1.0.4
mypy-extensions==0.4.3
pathspec==0.10.1
Pillow==9.3.0
platformdirs==2.5.3
proto-plus==1.22.1
protobuf==4.21.9
psycopg2-binary==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.6.0
pyparsing==3.0.9
pytz==2022.6
requests==2.28.1
rsa==4.9
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
tzdata==2022.6
uritemplate==4.1.1
urllib3==1.26.12
whitenoise==6.2.0
